+++ R/ifthen.R
@@ -1,16 +1,12 @@
-#' ifthen
-#'
-#' @description ifthen(a, b, c, d, e, f, ...) == if (a) b else if (c) d else if
-#' (e) f
-#'
-#' @param ... pairs of checks and corresponding return values
-#'
-#' @return
 #' @export
-#'
+#' @name ifthen
+#' @title Sum of vector elements
+#' @description `ifthen(a, b, c, d, e, f, ...)` == `if (a) b else if (c) d else
+#' if (e) f`
+#' @param ... pairs of checks and corresponding return values
 #' @examples
 #' x <- 2; y <- 2; z <- 1
-#' ifthen(x == 0, "foo", y == 0, "bar", z == 1, "this string gets returned")
+#' ifthen(x==0, "foo", y==0, "bar", z==1, "this string gets returned")
 ifthen <- function(...) {
     kwargs <- list(...)
     n <- length(kwargs)