+++ R/rm_all.R
@@ -1,12 +1,9 @@
-#' rmall
-#'
-#' @description Same as rm(list=ls())
-#'
-#' @return
 #' @export
-#'
-#' @examples \donttest{rm_all()}
+#' @name rm_all
+#' @title Remove all objects from global environment
+#' @description Same as rm(list=ls())
+#' @examples \dontrun{rm_all()}
 rm_all <- function() {
     e <- globalenv()
-    rm(list = ls(envir = e), envir = e)
+    rm(list=ls(envir=e), envir=e)
 }