+++ submit_to_cran.R
@@ -1,2 +1,0 @@
-# https://r-pkgs.org/release.html
-devtools::build(pkg=".", path=".")